Add files via upload
This code is for predicting stock price based on past experiences (rise and fall in stocks) using machine learning sklearn library and plotting the graphs of actual and predicted data with the help of matplotlib library of python. The calculations are being done using numpy library and the data set is being read by pandas library of python.